+++ sematic/ui/packages/main/src/types/Types.tsx
@@ -1,6 +1,11 @@
 import { AnyTypeRepr } from "@sematic/common/src/types";
 import BoolValueView from "src/types/boolean";
-import { CommonValueViewProps, TypeComponents, ValueView, ValueViewProps } from "src/types/common";
+import {
+  CommonValueViewProps,
+  TypeComponents,
+  ValueView,
+  ValueViewProps,
+} from "src/types/common";
 import DataclassValueView from "src/types/dataclass";
 import DataFrameValueView from "src/types/dataframetable";
 import DatetimeValueView from "src/types/datatime";
@@ -11,7 +16,6 @@ import ImageValueView from "src/types/image";
 import IntValueView from "src/types/int";
 import LinkValueView from "src/types/link";
 import ListValueView from "src/types/list";
-import MatplotlibFigureValueView from "src/types/matplot";
 import NoneValueView from "src/types/none";
 import PlotlyFigureValueView from "src/types/plotly";
 import StrValueView from "src/types/str";
@@ -20,11 +24,16 @@ import TupleValueView from "src/types/tuple";
 import { S3BucketValueView, S3LocationValueView } from "src/types/aws";
 export { renderSummary } from "src/types/common";
 
-
 // ValueComponent props
-type GenerateValueViewProps<U> = U extends AnyTypeRepr ? ValueViewProps<U> : never;
-type AllValueViewProps = CommonValueViewProps | GenerateValueViewProps<AnyTypeRepr>;
-type ExpandViewFunc<U> = U extends AllValueViewProps ? (props: U) => JSX.Element : never;
+type GenerateValueViewProps<U> = U extends AnyTypeRepr
+  ? ValueViewProps<U>
+  : never;
+type AllValueViewProps =
+  | CommonValueViewProps
+  | GenerateValueViewProps<AnyTypeRepr>;
+type ExpandViewFunc<U> = U extends AllValueViewProps
+  ? (props: U) => JSX.Element
+  : never;
 
 export type ComponentRenderDetails = {
   value: ExpandViewFunc<AllValueViewProps>;
@@ -50,14 +59,7 @@ const meta: Array<[string, ComponentRenderDetails]> = [
     "torch.utils.data.dataloader.DataLoader",
     { value: TorchDataLoaderValueView },
   ],
-  [
-    "plotly.graph_objs._figure.Figure",
-    { value: PlotlyFigureValueView },
-  ],
-  [
-    "matplotlib.figure.Figure",
-    { value: MatplotlibFigureValueView },
-  ],
+  ["plotly.graph_objs._figure.Figure", { value: PlotlyFigureValueView }],
   [
     "pandas.core.frame.DataFrame",
     {
@@ -68,24 +70,22 @@ const meta: Array<[string, ComponentRenderDetails]> = [
     "sematic.types.types.aws.s3.S3Bucket",
     {
       value: S3BucketValueView,
-    }
+    },
   ],
   [
     "sematic.types.types.aws.s3.S3Location",
     {
       value: S3LocationValueView,
-    }
+    },
   ],
   [
     "sematic.types.types.image.Image",
     {
       value: ImageValueView,
-    }
-  ]
+    },
+  ],
 ];
-
 
 meta.forEach(([key, value]) => {
   TypeComponents.set(key, value);
 });
-